Turned qudi.util.fit_models into a PEP420 namespace package in order to allow plugins to add fit models. Made use of the new helper function qudi.util.helpers.iter_modules_recursive in order to collect and import all available fit model objects from the namespace package.
