+++ qudi/util/datafitting.py
@@ -20,21 +20,47 @@ Copyright (c) the Qudi Developers. See the COPYRIGHT.txt file at the
 top-level directory of this distribution and at <https://github.com/Ulm-IQO/qudi/>
 """
 
+__all__ = ('is_fit_model', 'get_all_fit_models', 'FitConfiguration', 'FitConfigurationsModel',
+           'FitContainer')
+
+import importlib
+import logging
 import inspect
 import lmfit
 import numpy as np
 from PySide2 import QtCore
-from .mutex import Mutex
-from .units import create_formatted_output
-from . import fit_models as __models
+
+import qudi.util.fit_models as _fit_models_ns
+from qudi.util.mutex import Mutex
+from qudi.util.units import create_formatted_output
+from qudi.util.helpers import iter_modules_recursive
+from qudi.util.fit_models.model import FitModelBase
+
+
+_log = logging.getLogger(__name__)
 
 
 def is_fit_model(cls):
-    return inspect.isclass(cls) and issubclass(cls, __models.FitModelBase) and (
-                cls is not __models.FitModelBase)
+    return inspect.isclass(cls) and issubclass(cls, FitModelBase) and (cls is not FitModelBase)
 
 
-_fit_models = {name: cls for name, cls in inspect.getmembers(__models, is_fit_model)}
+# Upon import of this module the global attribute _fit_models is initialized with a dict
+# containing all importable fit model objects with names as keys.
+_fit_models = dict()
+for mod_finder in iter_modules_recursive(_fit_models_ns.__path__, _fit_models_ns.__name__ + '.'):
+    try:
+        _fit_models.update(
+            {name: cls for name, cls in
+             inspect.getmembers(importlib.import_module(mod_finder.name), is_fit_model)}
+        )
+    except:
+        _log.exception(
+            f'Exception while importing qudi.util.fit_models sub-module "{mod_finder.name}":'
+        )
+
+
+def get_all_fit_models():
+    return _fit_models.copy()
 
 
 class FitConfiguration:
