+++ qudi/util/fit_models/lorentzian.py
@@ -25,10 +25,10 @@ __all__ = (
 )
 
 import numpy as np
-from ._general import FitModelBase, estimator, correct_offset_histogram, smooth_data
-from ._general import sort_check_data
-from ._peak_helpers import estimate_double_peaks, estimate_triple_peaks
-from .linear import Linear
+from qudi.util.fit_models.model import FitModelBase, estimator
+from qudi.util.fit_models.helpers import correct_offset_histogram, smooth_data, sort_check_data
+from qudi.util.fit_models.helpers import estimate_double_peaks, estimate_triple_peaks
+from qudi.util.fit_models.linear import Linear
 
 
 def multiple_lorentzian(x, centers, sigmas, amplitudes):
